Split sidebar automation groups
